# service/src/docs/openapi.yaml
openapi: 3.0.1
info:
  title: Mage API
  description: Mage API
  contact:
    name: Mage Support
    email: mage@nga.mil
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0
  version: 6.0.1
servers: []
security:
  - mageToken: []
...
paths:
...
  /auth/{authenticationStrategy}/signin:
    post:
      tags: [ Authentication & Authorization ]
      summary: Authentication
      description: Authenticate a user using a given strategy.
      operationId: authenticate
      security:
        []
      parameters:
        - name: authenticationStrategy
          in: path
          description: Authentication strategy.
          required: true
          schema:
            type: string
            enum: [ local, ldap, saml, oauth, openidconnect ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UserIn' }
          multipart/form-data:
            schema: { $ref: '#/components/schemas/UserIn' }
      responses:
        200:
          description: Successfully signed in.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Authentication' }
  /auth/token:
    post:
      tags: [ Authentication & Authorization  ]
      summary: Device Authorization
      description: Authorize a device.  If device does not exist, it may also be created via this method.
      operationId: authorize
      security:
        - authenticationToken: []
      parameters:
        - name: uid
          in: query
          description: Id of the device that requires provisioning
          required: true
          schema: { $ref: '#/components/schemas/Device/properties/id' }
        - name: name
          in: query
          description: Name of the device
          schema: { $ref: '#/components/schemas/Device/properties/name' }
        - name: description
          in: query
          description: Description of the device
          schema: { $ref: '#/components/schemas/Device/properties/description' }
        - name: appVersion
          in: query
          description: Application version of the device (required if creating a new device)
          schema: { $ref: '#/components/schemas/Device/properties/appVersion' }
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DeviceIn' }
          multipart/form-data:
            schema: { $ref: '#/components/schemas/DeviceIn' }
      responses:
        200:
          description: Successful device authorization, Mage api token is returned.
        401:
          description:  User is not authorized.  Typically this means you need to authenticate `/auth/local/signin`.
        403:
          description: Failure to provision the device.  This typically means the device needs to be registered.
...
  /api/users/{userId}:
    parameters:
      - { $ref: '#/components/parameters/userIdInPath' }
    get:
      tags:
      - User
      description: >
        Return the user document whose `id` equals the `userId` path parameter.
        The client must have `READ_USER` permission.
      operationId: getUserById
      responses:
        200:
          description: The updated user document
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
    put:
      tags: [ User ]
      description: >
        Update the user document whose `id` equals the `userId` path parameter.
        The client must have `UPDATE_USER` permission.
      operationId: updateUserById
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UserIn' }
          multipart/form-data:
            schema: { $ref: '#/components/schemas/UserIn' }
      responses:
        200:
          description: user update response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
    delete:
      tags: [ User ]
      description: >
        Delete the user document whose `id` equals the `userId` path parameter.
        The client must have `DELETE_USER` permission.
      operationId: deleteUser
      responses:
        204:
          description: Success - user deleted
...
  /api/teams:
    post:
      tags: [ Team ]
      description: >
        Save a new team document to the database.  The requesting user must
        have `CREATE_TEAM` permission.
      operationId: createTeam
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TeamCreate' }
          application/x-www-form-urlencoded:
            schema: { $ref: '#/components/schemas/TeamCreate' }
      responses:
        200:
          description: Success - the created team document
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Team'
    get:
      tags: [ Team ]
      description: >
        Return all the teams.  The requesting client must have `READ_TEAM`
        permission.
      operationId: getTeams
      parameters:
        - in: query
          name: start
          description: Used with pagination.  Indicates the record number to start (leave blank for initial set of results).
          schema:
            type: integer
        - in: query
          name: limit
          description: Used with pagination.  Maximum number of records to return.
          schema:
            type: integer
        - in: query
          name: sort
          description: Describes the sort order.  Values allowed are asc, desc, ascending, descending, 1, and -1.
          content:
            application/json:
              schema:
                type: object
                required:
                  - name
                  - _id
                properties:
                  name:
                    type: string
                  _id:
                    type: string
        - in: query
          name: omit_event_teams
          description: >
            Whether to omit the event's implicit team that Mage automatically
            creates for adding individual users to an event.
          schema:
            type: boolean
        - in: query
          name: with_members
          description: >
            Return only teams that have the given user IDs as members.
          schema:
            type: array
            items:
              $ref: '#/components/schemas/User/properties/id'
        - in: query
          name: without_members
          description: >
            Return only teams that do not have the given user IDs as members.
          schema:
            type: array
            items:
              $ref: '#/components/schemas/User/properties/id'
        - in: query
          name: term
          description: >
            Return teams whose name or description field contains the given search term.
          schema:
            type: string
      responses:
        200:
          description: Success - an array of team documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Team'
...
  /api/teams/{teamId}:
    parameters:
      - $ref: '#/components/parameters/teamIdInPath'
    get:
      tags: [ Team ]
      description: >
        Return the team document whose ID matches the team ID path parameter.
        The requesting client must have `READ_TEAM` permission.
      operationId: getTeamById
      responses:
        200:
          description: Success - a team document
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Team'
    put:
      tags: [ Team ]
      description: >
        Update the team document whose ID matches the team ID in the path.
        The requesting user must have `UPDATE_TEAM` permission.
      operationId: updateTeamById
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TeamIn' }
          application/x-www-form-urlencoded:
            schema: { $ref: '#/components/schemas/TeamIn' }
      responses:
        200:
          description: Success - return the updated team document.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Team'
    delete:
      tags: [ Team ]
      description: >
        Delete the team document whose ID matches the team ID in the path.  The
        request user must have `DELETE_TEAM` permission.
      operationId: deleteTeam
      responses:
        204:
          description: Success - team deleted
  /api/teams/{teamId}/users:
    parameters:
      - { $ref: '#/components/parameters/teamIdInPath' }
    post:
      tags: [ Team ]
      operationId: addUserToTeam
      description: >
        Add a user to a team.  The request user must have `UPDATE_TEAM`
        permission as well as an ACL entry in the team document with `update`
        permission.
      requestBody:
        description: The user ID reference to add to the target team
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRef'
        required: true
      responses:
        200:
          description: Success - return the updated team document.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Team' }
  /api/teams/{teamId}/users/{userId}:
    parameters:
      - { $ref: '#/components/parameters/teamIdInPath' }
      - { $ref: '#/components/parameters/userIdInPath' }
    delete:
      tags: [ Team ]
      operationId: removeUserFromTeam
      description: >
        Remove a user from a team.  The requesting user must have `UPDATE_TEAM`
        permission as well as an ACL entry in the team document with `update`
        permission.
      responses:
        200:
          description: Success - return the updated team document.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Team' }
  /api/teams/{teamId}/acl/{userId}:
    parameters:
      - { $ref: '#/components/parameters/teamIdInPath' }
      - { $ref: '#/components/parameters/userIdInPath' }
    put:
      tags: [ Team ]
      operationId: setUserAccessForTeam
      description: >
        Update a team ACL entry.  The requesting user must have `UPDATE_TEAM`
        permission, as well as an ACL entry in the team document with `update`
        permission.
      requestBody:
        $ref: '#/components/requestBodies/ACLRoleUpdate'
      responses:
        200:
          description: Success - return the updated team document.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Team' }
    delete:
      tags: [ Team ]
      operationId: removeUserAccessFromTeam
      description: >
        Delete a team ACL entry.  The requesting user must have `UPDATE_TEAM`
        permission, as well as an ACL entry in the team document with `update`
        permission.
      responses:
        200:
          description: Success - return the updated team document.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Team' }
...
components:
  securitySchemes:
    authenticationToken:
      type: http
      description: >
        Token generated after successful authentication.  The client will exchange
        this token with an approved device for and API token.
      scheme: bearer
      bearerFormat: JWT
    mageToken:
      type: http
      description: >
        Mage generates an API token when a client authenticates
        initially.  The client can use the token for subsequent requests until
        it expires.
      scheme: bearer
  schemas:
...
    ColorHex:
      description: >
        This is a 3 or 4 byte hexadecimal string prefixed with '#' representing
        an RGB or RGBA color value, e.g. '#ff4545',
      type: string
      pattern: '#([a-fA-F0-9]{2}){3,4}'
...
    User:
      description: >
        The `User` schema defines the structure of persisted user documents as
        they exist in the database and as the server returns them in various
        responses.
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        displayName:
          type: string
        status:
          type: string
        email:
          type: string
          format: email
        createdAt:
          type: string
          format: date-time
        lastUpdated:
          type: string
          format: date-time
        active:
          type: boolean
        enabled:
          type: boolean
        avatarUrl:
          type: string
        iconUrl:
          type: string
        authentication:
          $ref: '#/components/schemas/AuthStatus'
        role:
          $ref: '#/components/schemas/Role'
        roleId:
          $ref: '#/components/schemas/Role/properties/id'
        icon:
          $ref: '#/components/schemas/UserIcon'
        phones:
          type: array
          items:
            $ref: '#/components/schemas/Phone'
        recentEventIds:
          type: array
          items:
            $ref: '#/components/schemas/Event/properties/id'
      oneOf:
        - { type: object, required: [ role ], not: { type: object, required: [ roleId ] } }
        - { type: object, required: [ roleId ], not: { type: object, required: [ role ] } }
        - { type: object, not: { type: object, required: [ role, roleId ] } }
    UserRef:
      description: >
        `UserRef` is an object that references an existing user document with
        the required `id` property.  Mage ignores any properties other than
        `id`.
      type: object
      allOf:
        - { $ref: '#/components/schemas/User' }
        - { type: object, required: [ id ] }
    UserIn:
      type: object
      properties:
        username: { $ref: '#/components/schemas/User/properties/username' }
        displayName: { $ref: '#/components/schemas/User/properties/displayName' }
        email: { $ref: '#/components/schemas/User/properties/email' }
        phone:
          type: string
          format: phone
        roleId: { $ref: '#/components/schemas/User/properties/roleId' }
        iconMetadata:
          type: object
          properties:
            type: { $ref: '#/components/schemas/UserIcon/properties/type' }
            color: { $ref: '#/components/schemas/UserIcon/properties/color' }
            text: { $ref: '#/components/schemas/UserIcon/properties/text' }
        icon:
          description: >
            This key is valid only when encoded as `multipart/form-data`.
            This is a binary image file attachment.
          type: string
          format: binary
        avatar:
          description: >
            This key is valid only when encoded as `multipart/form-data`.
            This is a binary image file attachment.
          type: string
          format: binary
      allOf:
        - { $ref: '#/components/schemas/UserPasswordPair' }
...
    UserIcon:
      description: >
        A user icon is the small image that appears on a map to show the user's
        location.
      type: object
      properties:
        type:
          description: The origin of the icon
          type: string
          enum:
            - create
            - upload
            - none
        color:
          description: Color is only applicable for `create` type icons.
          $ref: '#/components/schemas/ColorHex'
        text:
          description: >
            Two-letter text label that appears on the map icon; only applies to
            `create` type icons
          type: string
        contentType:
          description: The MIME type of the icon image
          type: string
        size:
          description: The image size in bytes
          type: integer
    UserPasswordPair:
      type: object
      properties:
        password:
          type: string
        passwordconfirm:
          type: string
      oneOf:
        - { type: object, required: [ password, passwordconfirm ] }
        - { type: object, not: { type: object, required: [ password, passwordconfirm ] } }
    Authentication:
      type: object
      required:
        - user
        - token
      properties:
        user:
          $ref: '#/components/schemas/UserRef'
        token:
          type: string
    AuthStatus:
      type: object
      required:
        - type
      properties:
        type:
          type: string
        security:
          type: object
          properties:
            locked:
              type: boolean
            lockedUntil:
              type: string
              format: date-time
            invalidLoginAttempts:
              type: number
            numberOfTimesLocked:
              type: number
    ACL:
      description: >
        The ACL (Access Control List) is a mapping of user IDs to the roles and
        permissions users have to access the parent resource of the ACL.
      type: object
      additionalProperties:
        description: >
          The keys of the ACL object are user ID strings.  The values specify
          the role and permissions that the user matching the associated key
          has to access a resource.
        type: object
        properties:
          role: { $ref: '#/components/schemas/ACLRole' }
          permissions:
            description: >
              The permissions of an ACL entry are currently a static mapping
              from the role.  The ACL entry includes the permissions associated
              with the role to be more explicit about what the role means and
              how the user of the ACL entry can access the resource.
            type: array
            items: { $ref: '#/components/schemas/ACLPermission' }
    ACLRole:
      type: string
      enum:
        - OWNER
        - MANAGER
        - GUEST
    ACLPermission:
      type: string
      enum:
        - read
        - update
        - delete
...
    DeviceIn:
      type: object
      required: [ uid ]
      properties:
        uid:
          type: string
        name:
          type: string
        description:
          type: string
        appVersion:
          type: string
        userId:
          $ref: '#/components/schemas/User/properties/id'
    Team:
      description: >
        A team is simply a grouping of users that commonly operate together.
        Teams facilitate adding groups of users to events.
      type: object
      properties:
        id:
          type: number
        name:
          type: string
        description:
          type: string
        users:
          type: array
          items:
            $ref: '#/components/schemas/UserRef'
        teamEventId:
          $ref: '#/components/schemas/Event/properties/id'
          description: >
            Mage adds this property to indicate that a team is the implicitly
            created team paired with an event for grouping the event's users
            on an individual basis.  This property references the ID of the
            event document to which the team belongs.  To add an individual
            user to an event, add the user to the team with the `teamEventId`
            property that references the target event.
        acl:
          $ref: '#/components/schemas/ACL'
    TeamIn:
      type: object
      properties:
        name: { $ref: '#/components/schemas/Team/properties/name' }
        description: { $ref: '#/components/schemas/Team/properties/description' }
        users:
          type: array
          items: { $ref: '#/components/schemas/UserRef' }
    TeamCreate:
      allOf:
        - $ref: '#/components/schemas/TeamIn'
        - { type: object, required: [ name ] }
...
    Phone:
      type: object
      properties:
        type:
          description: >
            The type of phone number currently defaults to `Main` and is
            static.
          type: string
        number:
          type: string
...
    Role:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        permissions:
          type: array
          items:
            type: string
      required:
        - id
        - name
        - description
        - permissions
...
  parameters:
    userIdInPath:
      in: path
      name: userId
      description: The ID of the target user document
      required: true
      example: 5d0b2bfeeec24262f1a5fdf3
      schema: { $ref: '#/components/schemas/User/properties/id' }
...
    teamIdInPath:
      in: path
      name: teamId
      description: The ID of the target team document
      required: true
      schema: { $ref: '#/components/schemas/Team/properties/id' }
...
  requestBodies:
...
    ACLRoleUpdate:
      required: true
      content:
        application/json:
          schema:
            type: object
            properties:
              role:
                $ref: '#/components/schemas/ACLRole'
            required: [ role ]